Corrected Anzahl of Nicht genehmigte Timesheets in Timesheet Overview
Issue: When Mitarbeiter is new (has no timesheets yet) and did not CREATE timesheet the past month(s),
the Anzahl of 'Nicht genehmigte' Timesheets shown in Timesheet Overview was null, until Vorgesetzte or Mitarbeiter
created first timesheet.
Now it shows correctly amount of 'Nicht genehmigte', no matter if timesheet was created or not.